LPS-174950 Make Action an abstract class and extract the script fields to the new ScriptAction model
