+++ modules/apps/portal-workflow/portal-workflow-kaleo-definition-api/src/main/java/com/liferay/portal/workflow/kaleo/definition/ScriptAction.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.workflow.kaleo.definition;
+
+import com.liferay.portal.workflow.kaleo.definition.exception.KaleoDefinitionValidationException;
+
+/**
+ * @author Rafael Praxedes
+ */
+public class ScriptAction extends Action {
+
+	public ScriptAction(
+			String name, String description, String executionType,
+			String script, String scriptLanguage, String scriptRequiredContexts,
+			int priority)
+		throws KaleoDefinitionValidationException {
+
+		super(ActionType.SCRIPT, name, description, executionType, priority);
+
+		_script = script;
+		_scriptLanguage = ScriptLanguage.parse(scriptLanguage);
+		_scriptRequiredContexts = scriptRequiredContexts;
+	}
+
+	public String getScript() {
+		return _script;
+	}
+
+	public ScriptLanguage getScriptLanguage() {
+		return _scriptLanguage;
+	}
+
+	public String getScriptRequiredContexts() {
+		return _scriptRequiredContexts;
+	}
+
+	private final String _script;
+	private final ScriptLanguage _scriptLanguage;
+	private final String _scriptRequiredContexts;
+
+}
